+++ frontend-opal/src/thunks/SearchResult.js
